reinstalled dotenv and then shifted api key and port to dotenv to resolve exposed api key issue
