+++ backend/.env
@@ -1,0 +1,2 @@
+GENERATIVE_AI_API_KEY="AIzaSyC7uzzgMxRA0JTFJsSE3QHdIa1j75Dc9_Q"
+PORT=3000
