--- backend/src/.env
@@ -1,2 +1,0 @@
-GENERATIVE_AI_API_KEY = "AIzaSyBC16h2iWOSrNSRWgkezrOo6ayTwZuDH7s";
-PORT = 3000;
