--- backend/src/controllers/.gitkeep
@@ -1,2 +1,0 @@
-Git keep has been added to each folder as Git doesn't track empty folders
-It's just a placeholder
